First version to move the processor mixin within the core.processor python module.
